+++ src/commands.rs
@@ -1,7 +1,9 @@
 mod convert;
 mod down;
+mod pause;
 mod rm;
 mod stop;
+mod unpause;
 mod up;
 mod version;
 
@@ -16,6 +18,8 @@ pub(crate) enum Command {
     Down(down::Args),
     Rm(rm::Args),
     Stop(stop::Args),
+    Pause(pause::Args),
+    Unpause(unpause::Args),
     Convert(convert::Args),
     Version(version::Args),
 }
@@ -26,6 +30,8 @@ pub(crate) async fn run(command: Command, config: Config) -> Result<()> {
         Command::Down(args) => down::run(args, config).await?,
         Command::Rm(args) => rm::run(args, config).await?,
         Command::Stop(args) => stop::run(args, config).await?,
+        Command::Pause(args) => pause::run(args, config).await?,
+        Command::Unpause(args) => unpause::run(args, config).await?,
         Command::Convert(args) => convert::run(args, config)?,
         Command::Version(args) => version::run(args),
     };