+++ src/podman/types.rs
@@ -9,7 +9,7 @@ with_prefix!(prefix_io_podman_compose "io.podman.compose.");
 #[serde(rename_all = "PascalCase")]
 pub(crate) struct Container {
     #[serde(with = "prefix_io_podman_compose")]
-    pub(crate) labels: ContainerLabels,
+    pub(crate) labels: Option<ContainerLabels>,
     pub(crate) names: VecDeque<String>,
 }